--- a/src/PushStream6.PerformanceTest/PushStream6.PerformanceTest.fsproj
+++ b/src/PushStream6.PerformanceTest/PushStream6.PerformanceTest.fsproj
@@ -12,6 +12,7 @@
 
   <ItemGroup>
     <PackageReference Include="BenchmarkDotNet" Version="0.13.1" />
+    <PackageReference Include="Cistern.ValueLinq" Version="0.1.14" />
   </ItemGroup>
 
   <ItemGroup>