@manofstick 2021-11-16
On a modern machine, we see the expected reuslt:

```
BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19042.1348 (20H2/October2020Update)
AMD Ryzen 9 3900X, 1 CPU, 24 logical and 12 physical cores
.NET SDK=6.0.100
  [Host]    : .NET 6.0.0 (6.0.21.52210), X64 RyuJIT DEBUG
  RyuJitX64 : .NET 6.0.0 (6.0.21.52210), X64 RyuJIT

Job=RyuJitX64  Jit=RyuJit  Platform=X64

|             Method |       Mean |     Error |    StdDev | Ratio | RatioSD |   Gen 0 |  Gen 1 | Allocated |
|------------------- |-----------:|----------:|----------:|------:|--------:|--------:|-------:|----------:|
|           Baseline |   4.151 us | 0.0813 us | 0.1567 us |  1.00 |    0.00 |       - |      - |         - |
|               Linq | 116.443 us | 0.4997 us | 0.4172 us | 28.29 |    0.83 |       - |      - |     400 B |
|          ValueLinq |  70.070 us | 1.1393 us | 1.0657 us | 16.91 |    0.57 |       - |      - |     192 B |
|      ValueLinqFast |  10.479 us | 0.0159 us | 0.0141 us |  2.54 |    0.09 |       - |      - |         - |
|              Array |  77.699 us | 0.7267 us | 0.6797 us | 18.75 |    0.69 | 32.3486 | 6.3477 | 272,864 B |
|                Seq | 230.210 us | 0.1567 us | 0.1389 us | 55.70 |    1.88 |       - |      - |     480 B |
|         PushStream |  22.800 us | 0.2386 us | 0.2232 us |  5.50 |    0.20 |       - |      - |     168 B |
|   FasterPushStream |   5.204 us | 0.1010 us | 0.1542 us |  1.26 |    0.06 |       - |      - |         - |
|       PushStreamV2 | 136.635 us | 0.4060 us | 0.3798 us | 32.97 |    1.11 |       - |      - |     216 B |
| FasterPushStreamV2 |   5.288 us | 0.1058 us | 0.1583 us |  1.28 |    0.06 |       - |      - |         - |
```