@manofstick 2021-11-17
More comparison (with the original evens checker); run on daughters laptop 

```
BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19043.1348 (21H1/May2021Update)
Intel Core i5-4300M CPU 2.60GHz (Haswell), 1 CPU, 4 logical and 2 physical cores
.NET SDK=6.0.100
  [Host]    : .NET 6.0.0 (6.0.21.52210), X64 RyuJIT DEBUG
  RyuJitX64 : .NET 6.0.0 (6.0.21.52210), X64 RyuJIT

Job=RyuJitX64  Jit=RyuJit  Platform=X64

|             Method |       Mean |     Error |    StdDev | Ratio | RatioSD |    Gen 0 | Allocated |
|------------------- |-----------:|----------:|----------:|------:|--------:|---------:|----------:|
|           Baseline |   4.692 us | 0.0393 us | 0.0349 us |  1.00 |    0.00 |        - |         - |
|               Linq | 155.707 us | 0.9607 us | 0.7500 us | 33.17 |    0.31 |   0.2441 |     400 B |
|          ValueLinq |  87.132 us | 0.6270 us | 0.4895 us | 18.56 |    0.20 |   0.1221 |     192 B |
|      ValueLinqFast |  21.925 us | 0.3047 us | 0.2850 us |  4.67 |    0.07 |        - |         - |
|              Array | 126.425 us | 0.9715 us | 0.8612 us | 26.95 |    0.24 | 172.3633 | 272,864 B |
|                Seq | 277.089 us | 2.0606 us | 1.9275 us | 59.11 |    0.60 |        - |     480 B |
|         PushStream |  37.415 us | 0.2456 us | 0.2298 us |  7.98 |    0.08 |   0.0610 |     168 B |
|   FasterPushStream |   7.214 us | 0.0818 us | 0.0765 us |  1.54 |    0.02 |        - |         - |
|       PushStreamV2 | 162.116 us | 0.7232 us | 0.6765 us | 34.55 |    0.29 |        - |     216 B |
| FasterPushStreamV2 |   7.275 us | 0.0988 us | 0.0876 us |  1.55 |    0.02 |        - |         - |

```